MARKER
The code works right now to stitch and save an image in tif.
It saves as the name itself.tif and also adds in image_shape.txt the dimension of the images.

The purpose of marker is as a reference of where the code works as now I am going to do heavy modifications to the code!

PS: Also forgot to say that i removed the ome extension as we don't use it