MARKER
The code works for the sample images I have. With align1 and align2 they get stitched. I then created align2 padded to have bigger images. And I ran the align segment of the code where it takes all 2 original and 4 padded images and align them with respect to align1 and it worked!

This is just a marker file to know that this version worked.
WARNING: the code to stitch is commented right now as I have already verified that it worked