Optimised memory for align
I have incorporated functions into one big get_align images. This is to reduce copying variables to another function and being able to delete them when not needed to save memory.

I have modified the code so that you do phase_cross_correlation with a subimage of a given size that is located at the center of it

I have added input lines in the code so that we can look at the ram to see the memory usage (to be removed later on)