Convert image to uint16
Reduced the image size by 4 as i have successfully converted float image to int.
What i did not know was the type of values we had (before did it with int16 and as it would go above 32k, the images would look weird. With uint16 it visually looks the same)