Corrected normalisation of downsizing
The function skimage.rescale brings the values of the image between 0-1 which visually seems like the image is fine but value wise we lose all the information. Thus i found out that we have an argument preserve_range which keeps the range of the values in our image instead of normalising them.
Thus rescaling was broken for analysis (not visually) but now works!