finalimage function & WARNING
Removed the option to remove dapi as it is integrated when you want ot save all channels in one image.
added an argument in main.py where you can ask to save final image. Final image is all channels - dapi channels except for the reference dapi.

WARNING: as i sort the list, the reference image is different from before. And it doesnt do the registration well.... this means that you need the right reference image for it to work OR the images are small and particular thus maybe it creates issue. Possibly no problem when taking bigger images??