Declare reference channel
Modified the code so that you need a csv file if you want to run your code.

And now you input the reference channel you want and it will do registration with that reference.
From my testing it seems to work.
But these can be dangerous modifications so need to verify constantly.

Added also a function that verifies if the channels for each image in CSV match those of image_shape. If not, CSV needs to be corrected.

I also cleaned a bit the code in czi.py