update get_img_dim
Instead of appending on .txt it overwrites. This removes the need to delete the txt file.

Maybe need to do that in stitching part for creation of image_shape....

VERY LIKELY: when you have a key error that comes from channel_check function it is often is due to image_txt file that has more channels than expected thus it tries to go over the number of indices in the different variables