Metadata and catch errors.
Added the metadata in the ome.tif file.
This was achieved by getting the metadata of a tif that had the information that i wanted. And then reversed engineered it to write the new metadata with the information that i want for each respective ome.tif.

Also added an argument resolution so that the user can put the scale he/she wants with respect to the microscope (NB: The scale gets modified automatically if you downscale)

Finally added some try functions to see if the user has the right CSV naming with respect to the files in the folder