Weird Qupath channel counting
For img 1 through 3 the channel numbers are fine in Qupath but when loading img 4-6, Qupath adds extra empty channels
This doesnt affect the data. The metadata looks fine. Tried on two different samples of czi and got the same result: img4 and 5 have 16 channels and img 6 have 9. I cannot find the reason why....