Corrected filepath calling
Encountered an issue where if the dataset is above 10, *.czi/tif would call the last one (instead of rnd1 it does rnd10) thus making the analysis incomplete and the marker names false.
I removed the general calling of * and now it is more specific. This means that i cannot do tifs anymore, they have to be ome.tif.