Added Timer and background
Added a timer for each major step in the code so that it can give an idea for users how long the code will take.

Also added the creation of the background.czi. The argument takes the filename and the code will reassemble and align (add channel name metadata only). The next step will be to do the subtraction for the final image.