Updated the FAQ and WARNING
Added the question if you have aicspylibczi that you cannot find.

ALSO
we are saving the image currently with tif.save and it is a function that is deprecated.
Now it is tif.write
Currently some systems download the old version still so I am keeping save as it still works but might need to change it in the future!