Readme update and image_processing step changed
Corrected a small error in Readme and removed the fact that the pyramidal step was inside the final_image step
Like this if you forgot the argument pyramidal you can disable all previous steps and run just pyramidal step. Also added the time to do the processing in pyramidal

Tried to run it on a big image, and qupath isn't able to upload it. gonna try without the compression..