MARKER FINAL PUSH
This commit is to have a marker where I know this code works with all its functions (tested on a tumor image with 49 channels 13'500x14'000 pixels). All parts of the code was able to run and produce the wanted output. This will be where I merge the branch to master so that master has the latest working version.
Very likely my last push. It has been a pleasure and I hope that this code will be developed further and be helpful for you all